refactor: move data services to providers
