+++ studio/src/app/components/dashboard/app-dashboard-deck-actions/app-dashboard-deck-actions.tsx
@@ -7,7 +7,7 @@ import i18n from '../../../stores/i18n.store';
 import {Deck} from '../../../models/data/deck';
 
 import {DeckDashboardCloneResult, DeckDashboardService} from '../../../services/deck/deck-dashboard.service';
-import {DeckService, getDeckService} from '../../../services/data/deck/deck.service';
+import {DeckService, getDeckService} from '../../../providers/data/deck/deck.service';
 
 import {AppIcon} from '../../core/app-icon/app-icon';
