+++ studio/src/app/pages/core/settings/app-templates/app-templates.tsx
@@ -12,7 +12,7 @@ import {AuthUser} from '../../../../models/auth/auth.user';
 
 import {signIn} from '../../../../utils/core/signin.utils';
 
-import {TemplateService} from '../../../../services/data/template/template.service';
+import {TemplateService} from '../../../../providers/data/template/template.service';
 import {renderI18n} from '../../../../utils/core/i18n.utils';
 import {modalController} from '../../../../utils/ionic/ionic.overlay';
