+++ studio/src/app/providers/data/slide/slide.firebase.service.ts
