+++ studio/src/app/providers/data/slide/slide.offline.service.ts
